backend/requirements-simple.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-multipart
python-dotenv
pydantic[email]
pydantic-settings
minio
redis
pypdf2
python-docx
openpyxl
python-pptx
beautifulsoup4
markdown
httpx
tenacity
loguru
aiofiles
python-jose[cryptography]
passlib[bcrypt]